Toggle bookmark of a post, list bookmarks of a post

+++ src/main/java/com/app/vibely/repositories/LikeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface LikeRepository extends JpaRepository<Like, Integer> {
-    void deleteAllByPostId(Long postId);
+    void deleteAllByPostId(Integer postId);
     Like findByPostIdAndUserId(Integer postId, Integer userId);
     Page<Like> findByPostId(Integer postId, Pageable pageable);
